+++ src/r4300/recomp.h
@@ -23,6 +23,7 @@
 #define M64P_R4300_RECOMP_H
 
 #include <stddef.h>
+#include <stdint.h>
 #if defined(__x86_64__)
   #include "x86_64/assemble_struct.h"
 #else
@@ -36,8 +37,8 @@ typedef struct _precomp_instr
      {
     struct
       {
-         long long int *rs;
-         long long int *rt;
+         int64_t *rs;
+         int64_t *rt;
          short immediate;
       } i;
     struct
@@ -46,9 +47,9 @@ typedef struct _precomp_instr
       } j;
     struct
       {
-         long long int *rs;
-         long long int *rt;
-         long long int *rd;
+         int64_t *rs;
+         int64_t *rt;
+         int64_t *rd;
          unsigned char sa;
          unsigned char nrd;
       } r;
